refactor(webapp): retire the host-command layer back-edge by relocating tray modules to base/
`shell/supplemental-commands/host-command.ts` statically imported 7 symbols
from `scoops/`, jumping three rungs up the layer stack (`fs/base → shell/git →
cdp → tools → core → scoops → ui`). It was the largest entry in
`layer-back-edge-baseline.json` — every other file was 4 or fewer.

Relocates rather than inverts. The issue proposed injecting the tray helpers
through setters registered from `ui/wc/wc-tray.ts`, mirroring
`setConnectedFollowersGetter`. That is realm-incorrect: `wc-tray.ts` runs only
on the page, while `host` executes in the standalone kernel worker, so the
setters would never fire and `host` / `join` / `leave` / `reset` would report
`inactive` while genuinely connected (the b268 symptom the shim readers were
added to fix).

Relocation avoids that entirely. Four of the six modules had no upward
dependency at all — they were low-layer code that merely lived in `scoops/`,
exactly the "pure helper in a higher-layer god module" disguise
`docs/review-patterns.md` § 10 names:

- `base/tray-follower-status.ts` — moved whole; imported only `base/tray-role.js`
- `base/tray-leader-status.ts`   — the self-contained registry slice extracted
                                   out of the 801-line manager, which stays put
- `base/tray-leave.ts` / `base/tray-join.ts` — moved whole; wire dispatchers
                                   that resolve an ambient transport and
                                   dispatch a CustomEvent or panel-RPC call,
                                   never touching an orchestrator
- `base/tray-url-config.ts`      — the pure URL/storage-key half split out of
                                   `tray-runtime-config.ts`; the fetch-bound
                                   half stays in `scoops/`
- `FloatType`                    — a bare string union, now in `base/tray-role.ts`

Per-realm behavior is preserved by construction: each module global moves
WHOLESALE and is re-exported, never re-declared, so it stays a single ESM
singleton. Whichever realm runs the managers sets it; every layer reads the
same one; the `localStorage` fallback moves verbatim. `scoops/` re-exports
every symbol under its established name, so the ~100 existing importers are
untouched — the same shape as `base/permissions-surface-registry.ts`.

Proof it is behavior-preserving: the multi-float harness added in #2558 passes
BYTE-IDENTICAL across this change. It drives the real defaults through real
setters and real globals with no `vi.mock`, so it is blind to which module a
symbol is declared in — which is exactly why it can prove the relocation.

`scoops/tray-join.ts` is deleted outright rather than left as a shim: its only
remaining caller was a test, and knip's production deadcode gate rejects a
src module no production code imports. The three tests whose source fully
relocated move to `tests/base/` to keep `tests/` mirroring `src/`.

Baseline: 44 back-edges in 21 files → 37 in 20. host-command.ts is gone from
the list, so the boy-scout gate holds the line at zero for that file.

Closes #2537
