+++ packages/dev-tools/tools/layer-back-edge-baseline.json
@@ -3,7 +3,6 @@
   "packages/webapp/src/fs/mount/backend-local.ts": 2,
   "packages/webapp/src/shell/bsh-watchdog.ts": 3,
   "packages/webapp/src/shell/supplemental-commands/hid-command.ts": 1,
-  "packages/webapp/src/shell/supplemental-commands/host-command.ts": 7,
   "packages/webapp/src/shell/supplemental-commands/playwright/handlers/interaction.ts": 1,
   "packages/webapp/src/shell/supplemental-commands/playwright/snapshot.ts": 3,
   "packages/webapp/src/shell/supplemental-commands/playwright/state.ts": 4,
