+++ packages/webapp/tests/base/tray-leave.test.ts
@@ -1,9 +1,6 @@
 import { describe, expect, it } from 'vitest';
-import { leaveTray, resolveAmbientLeaveTrayTransport } from '../../src/scoops/tray-leave.js';
-import {
-  TRAY_JOIN_STORAGE_KEY,
-  TRAY_WORKER_STORAGE_KEY,
-} from '../../src/scoops/tray-runtime-config.js';
+import { leaveTray, resolveAmbientLeaveTrayTransport } from '../../src/base/tray-leave.js';
+import { TRAY_JOIN_STORAGE_KEY, TRAY_WORKER_STORAGE_KEY } from '../../src/base/tray-url-config.js';
 
 interface FakeStorage {
   setItem(key: string, value: string): void;
